# feedback-app/openapi.yaml
openapi: 3.0.3
info:
  title: Feedback system
  description: Feedback and evaluation system for the studying courses.  # [CommonMark](http://commonmark.org/help/)
  version: 1.0
  contact:
    name: GitHub
    url: https://github.com/AlexPolarBear/feedback-system
servers:
  - url: http://127.0.0.1:5000
    description: The main server
paths:
...
  /tags:
    get:
      summary: Returns a list of all tags.
      responses:
        '200':
          description: A JSON array of tags
          content:
            application/json:
              schema:
                type: object
                items:
                  id:
                    type: integer
                  title:
                    type: string
                  type:
                    type: string
        '400':
          description: Unable to get a list of tags
        default:
          description: Unexpected error
    post:
      summary: Creates a tag.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                type:
                  type: string
      responses:
        '201':
          description: Tag successfully saved
        '400':
          description: Failed to add a tag
        default:
          description: Unexpected error
  /tags/{id}:
    get:
      summary: Returns a tag by id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type : integer
            minimum: 1
      responses: 
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                items:
                  id:
                    type: integer
                  title:
                    type: string
                  type:
                    type: string
        '422':
          description: There is no tag with this id
        default:
            description: Unexpected error
  /tags/delete/{id}:
    delete:
      summary: Delete a tag by id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type : integer
            minimum: 1
      responses: 
        '202':
          description: Tag successfully deleted
        '422':
          description: There is no score with this id
        default:
            description: Unexpected error
  /tags/update/type/{id}:
    post:
      summary: Updates a tag's type by id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
      responses:
        '201':
          description: Tag's type successfully updated
        '422':
          description: There is no tag with this id
        default:
          description: Unexpected error
  /tags/update/title/{id}:
    post:
      summary: Updates a tag's title by id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
      responses:
        '201':
          description: Tag's title successfully updated
        '422':
          description: There is no tag with this id
        default:
          description: Unexpected error